dataset_creation/requirements.txt
pandas>=1.3.0
biopython>=1.79
matplotlib>=3.4.0
numpy>=1.20.0
